Gun and It
Added the Gun
Added It

+++ The Dark/src/mygame/util/script/Script.java
@@ -21,7 +21,7 @@ public class Script {
     
     private Entity          entity;
     private ScriptManager   scriptManager;
-    private int             proximity;
+    private float           proximity;
     private boolean         inProx;
     private boolean         enteredProx;
     private Player          player;
